pyproject.toml
[tool.poetry]
name = "tests_lib"
version = "0.1.0"
description = ""
authors = ["Liubomyr Pavlovskyi <zzee19191919@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.4"
requests = "^2.32.3"
selenium = "^4.27.1"
python-dotenv = "^1.0.1"
webdriver-manager = "^4.0.2"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"